NetworkSession: allow discarding consecutive identical packets without decoding in certain cases (#6715)
This allows packet handlers to tell the network session to discard any following packets if their buffers were identical to the current one.

This is a very conservative and basic limiter. If we start having problems with multiple packets being spammed at once interleaved, this will become ineffective. However, for now, all the known spam bugs are of single packets, and the buffers of said packets are always identical. Dealing with interleaved packets would be quite a bit more complex. This mechanism is very simple and should avoid any negative side effects.